Avoid `sudo chown -R` on `${ANDROID_HOME}` on Linux
This seems to take several minutes, a decent fraction of the overall time in my
simple use.

Instead of changing the ownership of `${ANDROID_HOME}` instead arrange to run
individual commands under `sudo` when needed.

Note that `$PATH` is not preserved by `sudo` so we must use the full path to
the `sdkmanager`.

Another wrinkle is the cmdline-tools installation, since `tc.extractZip` and
`io.mv` do not include sudo-ish functionality. Instead precreate the target
directory with the ownership to allow the unpack as the current user.